Compiler no longer outputs strange characters in C strings: It properly escapes returns, tabs, \, ", etc, via the mangle-string-literal  function, and no longer depends on Proto's string output munging.
Added "assqn" (probably badly named) which is the same as assq, except
it takes an extra argument for the position in the list to check.

When writing characters, it now uses the appropriate long name for all
special characters by doing a lookup on $char-long-names instead of
just special casing #\space and #\newline.