Add a id-hash variant for tup's that recursively calls id-hash on all elements in the tup, multiply by 31 each time and adding. (A standard behavior used by the java collections API when computing hashcodes over the contained elements.)  This has been implemented after short debate with Chris Armstrong, and we both agree that no one loses by adding this base behavior; tups are immutable, and under the current scheme, if you lose the reference to your tup (and this is for keys of course), you can't get the damn thing back out again, which is stupid.
It's also a bit debatable why str-tab needs to be a separate class,
although str's are mutable, so I suppose that's okay.  But it certainly
makes sense that all immutable container objects should attempt to hash
for equivalency.

Or something.  Feel free to revert, but I think this is a good thing(tm).